chore: move getChargedExecutiveContent
